chore: moving files around, renamed plugins as loaders, no more workspace packages & plugins
