--- plugins/cpt-hcpcs/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "terminology-server-cpt-hcpcs",
-  "type": "module",
-  "devDependencies": {
-    "@types/bun": "latest",
-    "@types/node": "latest"
-  },
-  "peerDependencies": {
-    "typescript": "^5.0.0"
-  },
-  "dependencies": {
-    "@bonfhir/core": "^2.18.0",
-    "csv": "^6.3.8",
-    "xlsx": "https://cdn.sheetjs.com/xlsx-0.20.1/xlsx-0.20.1.tgz",
-    "yaml": "^2.3.4"
-  }
-}
