+++ server-setup/custom-loaders/rxnorm.ts
@@ -1,20 +1,24 @@
-import type { TerminologyPlugin } from "terminology-server-setup/terminology-plugin";
-import { codeSystem } from "./code-system";
 import { stringify } from "csv-stringify/sync";
 import { parse } from "csv-parse/sync";
-import type {
-  ConfigServer,
-  ConfigTaskEntry,
-} from "terminology-server-setup/configs";
-
-import {
-  packageCustomVocabulary,
-  unzipFiles,
-  uploadFiles,
-} from "terminology-server-setup/utils";
 import type { AuditEventOutcome } from "@bonfhir/core/r4b";
+import type { CodeSystem } from "@bonfhir/core/r4b";
 
-export default class RxNormPlugin implements TerminologyPlugin {
+import type { CustomLoader } from "./index";
+import type { ConfigServer, ConfigTaskEntry } from "@/configs";
+import { packageCustomVocabulary, unzipFiles, uploadFiles } from "@/utils";
+
+const codeSystem: Partial<CodeSystem> = {
+  resourceType: "CodeSystem",
+  url: "http://www.nlm.nih.gov/research/umls/rxnorm",
+  name: "RXNorm",
+  description:
+    "RxNorm is a normalized naming system for generic and branded drugs by the United States National Library of Medicine.",
+  status: "active",
+  publisher: "U.S. National Library of Medicine",
+  content: "not-present",
+};
+
+export default class RxNormLoader implements CustomLoader {
   name = "rxnorm";
   async uploadTerminology(
     server: ConfigServer,
