+++ supervisord.conf
@@ -22,8 +22,8 @@ redirect_stderr=true
 
 [program:server-bootstrap]
 user=bonfhir
-directory=/bonfhir
-command=bun start
+directory=/bonfhir/server-setup
+command=/bonfhir/server-setup/bootstrap.ts
 stdout_logfile=/dev/fd/1
 stdout_logfile_maxbytes=0
 redirect_stderr=true
